half it actually

+++ finetune.py
@@ -135,7 +135,7 @@ def main(num_epochs = 10, lr=0.00002):
 
     config = BloomConfig.from_pretrained(model_name)
     tokenizer = BloomTokenizerFast.from_pretrained(model_name, trust_remote_code=True)
-    model = BloomForCausalLM.from_pretrained(model_name, trust_remote_code=True).cuda() #do .half() for inference
+    model = BloomForCausalLM.from_pretrained(model_name, trust_remote_code=True).half().cuda() #do .half() for inference
     #model = model.quantize(4)
     
     thawed_params = freezer_bloom(model, 5)
